requirements.txt
--constraint=constraints.txt
brotli==1.1.0
cryptography==44.0.2
Cython==3.0.12
Flask==3.1.0
lockfile==0.12.2
lxml==5.3.1
mypy_extensions==1.0.0
numpy==1.23.5
openai==1.70.0
pandas==2.0.3
protobuf==3.20.1
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyOpenSSL==25.0.0
python-dotenv==1.1.0
tabulate==0.9.0
torch==2.4.1
whisperx==3.2.0
wincertstore==0.2.1; sys_platform == "win32"
zipp==3.17.0
gunicorn==20.1.0
